+++ Ma.Web/Data/ApplicationDbContext.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+﻿using Ma.Web.Models;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 
 namespace Ma.Web.Data
@@ -9,5 +10,7 @@ namespace Ma.Web.Data
             : base(options)
         {
         }
+
+        public DbSet<AppointmentItem> Appointments { get; set; }
     }
 }