+++ Ma.Web/Data/Migrations/00000000000000_CreateIdentitySchema.cs
@@ -4,7 +4,7 @@ using Microsoft.EntityFrameworkCore.Migrations;
 
 namespace Ma.Web.Data.Migrations
 {   
-    public partial class v1 : Migration
+    public partial class CreateIdentitySchema : Migration
     {
         protected override void Up(MigrationBuilder migrationBuilder)
         {