+++ Ma.Web/Models/AppointmentRepetitions.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace Ma.Web.Models
+{
+    /// <summary>
+    /// Values for appointment evaluation frequency.
+    /// </summary>
+    [Serializable]
+    public enum AppointmentRepetitions : byte
+    {        
+        Never   = 0,
+        Daily   = 1,
+        Weekly  = 2,
+        Monthly = 3,
+        Yearly  = 4
+    }
+}