+++ Ma.Web/Services/AppVersionService.cs
@@ -2,6 +2,9 @@
 
 namespace Ma.Web.Services
 {
+    /// <summary>
+    /// Represents a service to manage the version of app.
+    /// </summary>
     public interface IAppVersionService
     {
         string Version { get; }