+++ Ma.Web/Startup.cs
@@ -17,7 +17,7 @@ using System.IO;
 namespace Ma.Web
 {
     public class Startup
-    {       
+    {   
         public Startup(IConfiguration configuration)
         {
             Configuration = configuration;