+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log
+
+## [1.2] In process
+- Redesign main layout of footer.
+- Create new appointment table in database.
+
 ## [1.1] 2019-11-05
 - Redesign the main layout of the page (logo, menu).
 - Add version information to the main page.